Description:
The  out of memory issue primarily was caused because :

1. MimeTypeFinder is a connector class  used for MimeType detection .Before the
fix  a new instance of this class was created every time for each document 
discovered.Also this class uses (mime-util-2.1.1.jar)  jar  which has classes
whose objects were not being garbage collected as per jprobe analysis.As a
result the connector eventually ran out of memory. 

2. Also before the fix we were doing mimetype detection for all the documents in
consecutive database crawls irrespective whether they have changed or not.
Because of which  significant amount  of connector  time and memory  was
consumed  in computation of MimeType.

#Solution Approach

1.  Made the MimeTypeFinder class as singleton as result of which only a single
object of the class  is  created and used throughout for MimeType detection
which resulted into significant amount of drop in memory usage for mimetype
detection.

2.Also made MimeType detection only for changed documents when the
findProperty() method is called so that connectors time and memory is not wasted
for unchanged  documents in consecutive database crawls.

#Changes

1.Change in MimeTypefinder.java to make it a singleton class.
2.Changes in Util.java to remove mimetype detection code from setBlobContent
Method.
3.Changes in JsonDocument
  a.Implements the document interface .
  b.Addition of a instance level variable "changed" to determine whether a
document has changed .
4.Changes in DBClass
   a. set changed flag for modified documents