DB Fix MimeType detection for BLOBs/CLOBs.
Determine the mime type for the content while I have the
byte array of content in Util.setBinaryContent().  The 
previous mime type handling would fail to actually supply
the mime type property in the Document in time to be noticed.
(The set of property names would not contain PROPNAME_MIMETYPE,
so the CM whould never ask for it.)

This maintains the check for mime type support level in
the document findProperty() call to fetch the Content.
This allows the documents to added to the SnapshotRepository
so they don't keep appearing as new documents on every pass.

Deleted the redundant MimeTypeFinder, and just replaced its
usage with the MimeTypeDetector in the Connector Manager.

Code Review:
http://codereview.appspot.com/6506100