DO NOT INCLUDE THIS REVISION IN A RELEASE WITHOUT INCLUDING ALL OF THE BUG 7501553 FIXES THROUGH PART FOUR AT r427.
Fix bug 7301552: Reduce Snapshot Document Size and Fix LOB Checksums.

When indexing millions of documents, the snapshot store files achieve
gigabyte proportions. This impacts performance in multiple ways, most
notably in stitch().

The JsonObjectUtil maintains two sets of properties for the document
under construction: a set of SPI Values for the SPI Document, and
a separate set of properties in a JSON object that was a subset of the
SPI properties.  Specifically, the latter did not include the content.

The JSON object is used as the Diffing snapshot document for change
detection.  Since it contains the SHA-1 checksum for all of the 
properties and the content, the JSON object did not really need 
to contain most of the SPI properties for comparison, as a change
to a property would alter the checksum.

This change modifies the JSON object used for the Diffing snapshot
to contain only the DocId and the SHA-1 checksum of the other
properties and the content.  This reduces the size of the stringified
snapshot document to 55 chars, plus the length of the DocId.

This change also removes the SHA-1 checksum from the SPI properties
of the Document, since the checksum is used only by the Diffing
connector, and would just pollute the GSA word index with no benefit.

Finally, this change fixes the calculation of the checksum with
BLOB/CLOB content.  The properties and the LOB content are now
factored into the same message digest to calculate the sum.
Previusly, the code just checksummed the sum of the separate
checksums.

TODO: The separation between JsonObjectUtil, JsonDocumentUtil,
and Util(when it included setBinaryContent()) seems forced
and artificial.

Code Review:
http://codereview.appspot.com/6586043