DB Refactor DBConnectorType Part 2
This refactors DBConnectorType in anticipation of allowing better presentation
of the form fields. This change makes the config form construction thread-safe
by moving the ResourceBundle for localization into a instantiated ConfigForm
object for each config form that is constructed.

Although other clean up could happen, I postpone that because of the whitespace
changes incurred when adding an inner class.

TODO:
  - Add nicer display elements, esp. wrt LOB fields, BaseUrl, etc.
  - Improve layout
  
Code Review:
http://codereview.appspot.com/6560069