Fix a problem with DBSnapshotRepository introduced in r428. This change returns a new RepositoryHandlerIterator on each call to iterator in DBSnapshotRepository.
As part of the cleanup of RepositoryHandlerIterator in r428 I had fixed
another bug without fixing a third that depended on it.

The bug I fixed in r428 was that after RepositoryHandlerIterator.hasNext
returned false, it would return true the following time it was called and
iterate over the repository again. I did not notice that this behavior
was intertwined with an issue in DBSnapshotRepository where it has a
single instance of RepositoryHandlerIterator, and returns it each time
iterator is called. (I had added a TODO about that in r427.) So when the
iterator was exhaused the main loop in the diffing package would exit,
get a new SnapshotRepository iterator (really the same one as before),
and loop over the repository again.

After r428, however, the SnapshotRepository iterator is exhaused and
stays that way, as it should, but DBSnapshotRepository is still returning
the same iterator. So in the second pass the connector will delete
everything it indexed, and it won't reindex everything until the
connector is reset.

Code review: https://codereview.appspot.com/6873053/