Fix bug 7301553, part six: Errors in DocumentHandle serialization. Part six fixes the final problem, deserializing google:lastmodified correctly.
It is serialized as an ISO 8601 string, and was deserialized as a
string value. As a string, it wouldn't get formatted as an RFC 822
date for the feed, but left as an ISO 8601 string, which is ignored
by the GSA in that context.

The fix is simply to reconstruct the DateValue when deserialzing the
ISO 8601 string.

Other changes:

* Fixed a couple of leftover comments in the tests from r416 that were
  not reverted along with the code in r427.

Code review: https://codereview.appspot.com/7037044/