Fix Bug 7301515: Reduce memory use in iBATIS
The default MyBatis TypeHanders for BLOB and CLOB
fields return arrays of bytes or Strings, respectively.
This can consume considerable amounts of memory for 
a large number of LOBs returned in a query.

This change adds custom type handlers that move 
the LOB content into FileBackedOutputStream-backed 
ContentHolders for future consumption as each row
of the ResultSet is processed inside MyBatis.

Because this moves the use of FileBackedOutputStream
to a much lower level, I refactored the ContentHolder
code.  I added a DigestContentHolder that keeps the
MessageDigest used to determine the document checksum
open for updating until the checksum is requested.

Global custom typeholders required an upgrade to 
MyBatis 3, done in r439.

Code Review:
http://codereview.appspot.com/7125043