Fix Bug 7732455, part 3: Preserve Sort Order in Encoded DocIds - NULL sorting
This change fixes the sort order of NULL values in docid, using the
sort order of nulls in the database implementation as specified in the
DatabaseMetaData.  If the null sort order is not available from the
DatabaseMetaData, the user can override it by setting a nullsAreSortedLow
property in the connector's Advanced Configuration.

See also:
b/7928861: Avoid unordered snapshots in the snapshot files.

Code review:
http://codereview.appspot.com/7199049