Fix issue 10598470: Column names are case-sensitive.
All of the column name properties except the primary key were case and
whitespace-sensitive. This change updates the configuration map passed
into validateConfig with the correct case. I had originally just
looked up the correct column name at runtime, but that fails with the
configured type handler in the MyBatis SQL map (the correct case is
not available when the SQL map is written). Another rejected option
was to validate the configuration in DBContext, that is, during
connector instantiation. That might be nice, but it was too much
unnecessary work.

If validateConfig fails, the column name properties may be rewritten
in the correct case on the redisplayed form. It just depends on the
order of property validation.

If you manually edit the properties file, then you must enter the
fields in the correct case. However, I preserved the whitespace
trimming in DBContext, even though it's a duplicate effort, to avoid
silly errors.

I refrained from refactoring the code for missing attributes as much
as I did for invalid column names because that code will be affected
by part two of changing the radio buttons on the configuration form.

Code review: https://codereview.appspot.com/13504046/