Improve document content options on the configuration form, part one.
Part one rearranges the form, and adds a fourth radio button for the
implicit "noExt" or stylesheet option, and marks optionally required
fields, but does change the display or validation of the radio buttons
themselves.

If one of the three original radio buttons is selected, but the
corresponding fields are left empty, then the "notExt" or Stylesheet
option is implied, but not displayed as such. In part two, such a
configuration will be rejected if submitted, and the implied
configuration will be displayed on the populated form.

* Rename "External Metadata" to "Document Content", since the options
  really define where the content comes from.

* Remove the duplicate hidden "extMetadataType" field, which
  conflicted with the "extMetadataType" radio buttons.

* Remove the unused "externalMetadata" field and use "extMetadataType"
  as the field for the renamed "Document Content" label.

* Rearrange the configuration form to put all of the document content
  options together.

* Use JavaScript to dynamically mark fields as required (with a red
  star) when the corresponding radio button is selected.

* Use "disabled" rather than "readonly" for the unselected options.

* Add labels for the radio buttons so you can click them.

* Minor validation changes, since unselected options are disabled and
  therefore not submitted with the form, they may be missing (that is,
  null), which is OK.

Other changes:

* Renamed BLOB/CLOB Field to BLOB or CLOB Field, and Fetch URL Field
  to Display URL Field.

Code review: https://codereview.appspot.com/13424045/