Fix compiler errors against CM with Guava 11.
The database connector was using a copy of TestDirectoryManager from
the file system connector. The FS original was moving into CM and
later updated for Guava 11 (which removed the Files.deleteRecursively
method), but the DB copy was not updated.

The change here removes the local DB copy of TestDirectoryManager and
uses the one from CM.

CM fix for Guava 11:
https://github.com/googlegsa/manager.v3/commit/8bcb17da171eac1853ff1fe3e3e0ae94906fda31

Code review: https://codereview.appspot.com/229640043/